MixinPlayerMovementInput & Freecam no view bob when baritone goal process is active
