+++ src/main/resources/liquidbounce.mixins.json
@@ -7,6 +7,7 @@
   "mixinPriority": 1337,
   "client": [
     "authlib.MixinMinecraftClient",
+    "baritone.MixinPlayerMovementInput",
     "minecraft.block.MixinBlock",
     "minecraft.block.MixinBlockCollisionSpliterator",
     "minecraft.block.MixinBlockView",
